remove extensions
